fix broken routes
