+++ CHEMIST/templates/add_product.html
@@ -26,7 +26,7 @@
             <a class="nav-link" href="{{ url_for('index') }}">Home</a>
           </li>
           <li class="nav-item">
-            <a class="nav-link" href="{{ url_for('products') }}">View Products</a>
+            <a class="nav-link" href="{{ url_for('products..products') }}">View Products</a>
           </li>
         </ul>
       </div>
