+++ CHEMIST/templates/dashboard.html
@@ -24,10 +24,10 @@
           <a class="nav-link" href="{{ url_for('index') }}">Home</a>
         </li>
         <li class="nav-item">
-          <a class="nav-link" href="{{ url_for('logout') }}">Logout</a>
+          <a class="nav-link" href="{{ url_for('auth.logout') }}">Logout</a>
         </li>
         <li class="nav-item">
-          <a class="nav-link" href="{{ url_for('products') }}">Medicines</a>
+          <a class="nav-link" href="{{ url_for('products.products') }}">Medicines</a>
         </li>
       </ul>
     </div>
