+++ CHEMIST/templates/login.html
@@ -16,7 +16,7 @@
             <h3 class="text-center">Login</h3>
           </div>
           <div class="card-body">
-            <form method="POST" action="{{ url_for('login') }}">
+            <form method="POST" action="{{ url_for('auth.login') }}">
               {{ form.hidden_tag() }} <!-- Include CSRF token -->
               <div class="form-group">
                 {{ form.email.label(class="form-control-label") }}
@@ -36,7 +36,7 @@
             </form>
           </div>
           <div class="card-footer">
-            <p class="text-center">Don't have an account? <a href="{{ url_for('register') }}">Register</a></p>
+            <p class="text-center">Don't have an account? <a href="{{ url_for('auth.register') }}">Register</a></p>
           </div>
         </div>
       </div>
