+++ CHEMIST/templates/register.html
@@ -18,7 +18,7 @@
             <h3 class="text-center">Register</h3>
           </div>
           <div class="card-body">
-            <form method="POST" action="{{ url_for('register') }}">
+            <form method="POST" action="{{ url_for('auth.register') }}">
               {{ form.hidden_tag() }} <!-- Include CSRF token -->
               <div class="form-group">
                 <label for="name">Name</label>
@@ -38,7 +38,7 @@
               </div>
               <button type="submit" class="btn btn-primary btn-block">Register</button>
               <div class="card-footer">
-                <p class="text-center">Already have an account? <a href="{{ url_for('login') }}">Login</a></p>
+                <p class="text-center">Already have an account? <a href="{{ url_for('auth.login') }}">Login</a></p>
               </div>
             </form>
           </div>
